requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
imageio>=2.31
Pillow>=10.0
tqdm>=4.66
requests>=2.31
xgboost>=1.7
fastapi>=0.110
uvicorn>=0.29
pydantic>=2.6
streamlit>=1.32
plotly>=5.20
pytest>=7.4
PyYAML>=6.0
